One class per file - PSR 1
